Create code
In this example, the answers are stored in an array and the spinWheel function uses Math.random() to generate a random index that corresponds to an answer. The spinWheel function is called when the button with the id "spinButton" is clicked.
You can add your own variables to the answers array and add different styles to the wheel.